Consolidated diagnostics and tests
